backend/requirements.txt
# Core Django framework
Django==5.2.2

# Django REST Framework for API
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.4.0

# Form handling and UI components
django-crispy-forms==2.3
crispy-bootstrap5==2024.10

# Rich text editor
django-ckeditor-5==0.2.13

# Database drivers
psycopg2-binary==2.9.9

# Image processing
Pillow==10.4.0

# Environment variables
python-dotenv==1.0.1

# Production server
gunicorn==23.0.0

# Background tasks and caching
celery==5.4.0
redis==5.1.1

# API documentation
drf-yasg==1.21.8

# Additional useful packages for Django development
python-decouple==3.8
whitenoise==6.8.1

# For development only (optional)
# django-debug-toolbar==4.4.6
# django-extensions==3.2.3
